\name{.compareAllelicSumScoresPvalues}
\alias{.compareAllelicSumScoresPvalues}
\title{compute the H1 allelic sum score using both implementations...}
\usage{.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareAllelicSumScoresPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
}
\description{compute the H1 allelic sum score using both implementations}
\value{a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareAllelicSumScoresPvalues}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


}
\arguments{\item{gws}{a GWSified dataset of class \code{\link[GenABEL]{gwaa.data-class}}}
\item{nnbc_params}{the NNBC parameters, see \code{\link{parameters}}}
\item{phenotypes}{an optional vector of phenotypes}
\item{bins}{an optional dataframe of bins. if not given it will use a single bin
containing all the dataset}
}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

\alias{.compareAllelicSumScoresPvalues}

