\name{.compareDivariateLikelihood}
\alias{.compareDivariateLikelihood}
\title{compute the H1 DivariateLikelihood  using both implementations...}
\usage{.compareDivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareDivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareDivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareDivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
}
\description{compute the H1 DivariateLikelihood  using both implementations}
\value{a list of results\code{.compareDivariateLikelihood}: a list of results


\code{.compareDivariateLikelihood}: a list of results


\code{.compareDivariateLikelihood}: a list of results


\code{.compareDivariateLikelihood}: a list of results


\code{.compareDivariateLikelihood}: a list of results


\code{.compareDivariateLikelihood}: a list of results


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columnd: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareDivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


}
\arguments{\item{gws_bin}{a GWSified dataset of class \code{\link[GenABEL]{gwaa.data-class}} truncated to a single bin}
\item{nnbc_params}{the NNBC parameters, see \code{\link{parameters}}}
\item{phenotypes}{an optional vector of phenotypes}
\item{gws}{a GWSified dataset of class \code{\link[GenABEL]{gwaa.data-class}}}
\item{bins}{an optional dataframe of bins. if not given it will use a single bin
containing all the dataset}
}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

\alias{.compareDivariateLikelihood}

