\name{.compareDivariateLikelihoodPvalues}
\alias{.compareDivariateLikelihoodPvalues}
\title{compute the H1 DivariateLikelihood pvalues using both implementations...}
\usage{.compareDivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareDivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareDivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareDivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareDivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareDivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareDivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareDivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
}
\description{compute the H1 DivariateLikelihood pvalues using both implementations}
\value{a list of results\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


\code{.compareDivariateLikelihoodPvalues}: a list of results


}
\arguments{\item{gws_bin}{a GWSified dataset of class \code{\link[GenABEL]{gwaa.data-class}} truncated to a single bin}
\item{nnbc_params}{the NNBC parameters, see \code{\link{parameters}}}
\item{simple_permutations}{the number of permutations for the R simple implementation}
\item{phenotypes}{an optional vector of phenotypes}
\item{bins}{an optional dataframe of bins. if not given it will use a single bin
containing all the dataset}
\item{gws}{a GWSified dataset of class \code{\link[GenABEL]{gwaa.data-class}}}
}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

\alias{.compareDivariateLikelihoodPvalues}

