\name{.compareUnivariateLikelihood}
\alias{.compareUnivariateLikelihood}
\title{compute the H1 univariateLikelihood using both implementations...}
\usage{.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws_bin, nnbc_params, phenotypes=gws_bin@phdata$pop,
    covariables=data.frame())
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
.compareUnivariateLikelihood(gws, nnbc_params, phenotypes=gws@phdata$pop, covariables=data.frame(),
    bins)
}
\description{compute the H1 univariateLikelihood using both implementations}
\value{a list of results\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a list of results


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


\code{.compareUnivariateLikelihood}: a data frame with columns: bin chr start end nb_snps nb_permutations pv_divariate score_divariate fdr_divariate simple_divariate_llh


}
\arguments{\item{gws_bin}{a NNBCified dataset of class \code{\link[GenABEL]{gwaa.data-class}} truncated to a single bin}
\item{nnbc_params}{the NNBC parameters, see \code{\link{make.parameters}}}
\item{phenotypes}{an optional vector of phenotypes}
\item{gws}{a GWSified dataset of class \code{\link[GenABEL]{gwaa.data-class}}}
\item{bins}{an optional dataframe of bins. if not given it will use a single bin
containing all the dataset}
}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

\alias{.compareUnivariateLikelihood}

