\name{.compareUnivariateLikelihoodPvalues}
\alias{.compareUnivariateLikelihoodPvalues}
\title{compute the H1 univariateLikelihood pvalues using both implementations...}
\usage{.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws_bin, nnbc_params, simple_permutations,
    phenotypes=gws_bin@phdata$pop, covariables=data.frame())
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
.compareUnivariateLikelihoodPvalues(gws, nnbc_params, simple_permutations, phenotypes=gws@phdata$pop,
    covariables=data.frame(), bins)
}
\description{compute the H1 univariateLikelihood pvalues using both implementations}
\value{a list of results\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


\code{.compareUnivariateLikelihoodPvalues}: a list of results


}
\arguments{\item{gws_bin}{a NNBCified dataset of class \code{\link[GenABEL]{gwaa.data-class}} truncated to a single bin}
\item{nnbc_params}{the NNBC parameters, see \code{\link{make.parameters}}}
\item{simple_permutations}{the number of permutations for the R simple implementation}
\item{phenotypes}{an optional vector of phenotypes}
\item{gws}{a GWSified dataset of class \code{\link[GenABEL]{gwaa.data-class}}}
\item{bins}{an optional dataframe of bins. if not given it will use a single bin
containing all the dataset}
}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

\alias{.compareUnivariateLikelihoodPvalues}

