\name{concept#bin}
\alias{concept#bin}
\alias{concept#bin}
\title{Definition of bins}
\description{A bin is a genomic region, defined by its chromosome and is start and end positions.
It corresponds to a contiguous set of SNPs.}
\details{Notations:

B is the number of bins.
A given bin contains J SNPs.}
\keyword{concept}
\concept{bin}
\alias{bin}

