\name{concept#regularization.of.table.estimates}
\alias{concept#regularization.of.table.estimates}
\alias{concept#regularization.of.table.estimates}
\title{Regularizations of the estimates from contingency tables}
\description{To obtain more regular estimates, a constant is added to all cell counts.
It is a Dirichlet prior on parameters. This constant is
can be chosen to be \eqn{C=\alpha . \bar{n}}, where
\eqn{\alpha} is the chosen error rate and
\eqn{\bar{n}} is the mean number of individuals
per cell. This constant means that uncertainty on low cell counts
is high, not only because of randomness, but also because of
genotyping errors.}
\keyword{concept}
\concept{regularization of table estimates}
\alias{regularization_of_table_estimates}

