\name{gwsForBin}
\alias{gwsForBin}
\title{create a subset of a gws for a given bin...}
\usage{gwsForBin(gws, chr, start, end)
}
\description{create a subset of a gws for a given bin}
\value{a (sub) dataset as a \code{\link[GenABEL]{gwaa.data-class}} object, or NULL}
\arguments{\item{gws}{the dataset as a \code{\link[GenABEL]{gwaa.data-class}} object}
\item{chr}{the chromosome of the bin}
\item{start}{the start of the bin}
\item{end}{the end of the bin}
}

