\name{makeBinsForSnps}
\alias{makeBinsForSnps}
\title{make a bin for every SNP of the dataset...}
\usage{makeBinsForSnps(gws)
}
\description{make a bin for every SNP of the dataset}
\value{a sorted dataframe of bins (chr_name, start, end, bin_index)}
\arguments{\item{gws}{the dataset as a \code{\link[GenABEL]{gwaa.data-class}} object}
}

